fix: service worker duplicate configuration, updated react-router examples
