+++ examples/react-router/pwa-assets/tsconfig.json
@@ -10,7 +10,7 @@
       "~/*": ["./app/*"]
     },
     "resolveJsonModule": true,
-    "types": ["@vite-pwa/remix/react-router-sw", "vite/client", "vite-plugin-pwa/vanillajs"],
+    "types": ["@vite-pwa/remix/remix-sw", "vite/client", "vite-plugin-pwa/vanillajs"],
     "allowJs": true,
     "strict": true,
     // Vite takes care of building everything, not tsc.
