+++ examples/react-router/pwa-simple-sw/app/entry.client.tsx
@@ -4,7 +4,7 @@
  * For more information, see https://remix.run/file-conventions/entry.client
  */
 
-import { HydratedRouter } from 'react-router/dom';
+import { HydratedRouter } from 'react-router/dom'
 import { StrictMode, startTransition } from 'react'
 import { hydrateRoot } from 'react-dom/client'
